+++ custom_components/tmt_chow/__init__.py
@@ -20,8 +20,8 @@ from .const import (
     DOMAIN,
     PLATFORMS,
 )
-from .hub import TmtChowHub
 from .mqtt import MqttError
+from .ps21050d_hub import TmtChowHub
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool: